Update to gtk4

+++ Sources/SwiftGtk/Color.swift
@@ -1,16 +1,16 @@
 import CGtk
 
 public struct Color {
-    public var red: Double
-    public var green: Double
-    public var blue: Double
-    public var alpha: Double
+    public var red: Float
+    public var green: Float
+    public var blue: Float
+    public var alpha: Float
 
     public init(
-        _ red: Double,
-        _ green: Double,
-        _ blue: Double,
-        _ alpha: Double = 1
+        _ red: Float,
+        _ green: Float,
+        _ blue: Float,
+        _ alpha: Float = 1
     ) {
         self.red = red
         self.green = green
